refactor(categories): user can multiple categories
this refactor ensures best practice is followed, by taking away methods
from the controller and placing it in the model as a scope method. It
also increases test coverage for edge cases

[(Fixes|Delivers|Start|Finishes) #[PIVOTAL_TRACKER_ID]]
