+++ app/views/events/_form.html.erb
@@ -66,7 +66,7 @@
               <div class="row">
                 <div class="input-field col l7">
                   <i class="fa fa-th-large prefix icon-setting"></i>
-                  <%= f.collection_select :category_id, Category.all, :id, :name, { prompt: "Choose a category"} , class: "browser-default", blank_option: true, required: true%>
+                  <%= f.collection_select :category_id, Category.list, :id, :name, { prompt: "Choose a category"} , class: "browser-default", blank_option: true, required: true%>
                 </div>
 
                 <div class="input-field col l3">
