+++ spec/features/category_spec.rb
@@ -1,0 +1,10 @@
+require "rails_helper"
+
+RSpec.feature "Category:", type: :feature, js: true do
+  before do
+    @category = create(:category)
+  end
+
+  scenario "User visits subdomain" do
+  end
+end
