+++ spec/models/category_spec.rb
@@ -11,5 +11,4 @@ RSpec.describe Category, type: :model do
   it { should be_valid }
 
   it { should validate_presence_of :name }
-  it { should validate_uniqueness_of :name }
 end
