--- test/decorators/category_decorator_test.rb
@@ -1,4 +1,0 @@
-require "test_helper"
-
-class CategoryDecoratorTest < Draper::TestCase
-end
